asdadwadaw

+++ src/main/java/com/j23/server/services/restaurant/qna/QnaService.java
@@ -14,6 +14,8 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.UUID;
 
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
 @Service
 @Transactional
 @RequiredArgsConstructor
@@ -24,11 +26,12 @@ public class QnaService {
   public QnA addQna(QnA qnA) {
     if (qnaRepository.existsByQuestion(qnA.getQuestion())) {
       throw new ConflictException("Question already exists");
-    };
+    }
+    ;
 
     QnA lastQna = qnaRepository.findTopByOrderByCreatedOnDesc();
-    if(lastQna != null) {
-      qnA.setNumber(lastQna.getNumber() +1);
+    if (lastQna != null) {
+      qnA.setNumber(lastQna.getNumber() + 1);
     } else {
       qnA.setNumber(1);
     }
@@ -39,18 +42,23 @@ public class QnaService {
   public QnA updateQna(QnA qnA) {
     if (qnaRepository.existsByQuestionAndIdIsNot(qnA.getQuestion(), qnA.getId())) {
       throw new ConflictException("Question already exists");
-    };
+    }
+    ;
 
     return qnaRepository.save(qnA);
   }
 
   public Page<QnA> findAllQna(String search, Integer page, Integer size) {
+    if (search == null) {
+      return qnaRepository.findAll(PageRequest.of(page, size));
+    }
     return qnaRepository.findAllByQuestionContaining(search, PageRequest.of(page, size));
   }
 
   public void deleteQna(UUID id) {
-//    try {
-//      System.out.println("The uuid : " + id);
-      qnaRepository.deleteById(id);
+    if (!qnaRepository.existsById(id)) {
+      throw new ResponseStatusException(NOT_FOUND, "Qna not found");
+    }
+    qnaRepository.deleteById(id);
   }
 }
